pyproject.toml
[project]
name = "lightning-template-hydra"
version = "0.1.0"
description = "Aditya project"
readme = "README.md"
requires-python = ">=3.12"

# Dependencies required for the project
dependencies = [
    "black>=24.8.0",
    "lightning[extra]>=2.4.0",
    "pytest>=8.3.3",
    "pytest-cov>=4.1.0",  # Added for coverage reporting
    "rich>=13.8.1",
    "tensorboard>=2.17.1",
    "timm>=1.0.9",
    "torch>=2.4.1",
    "torchvision>=0.19.1",
    "kaggle>=1.5.12",
    "hydra-core>=1.3.2",  # Added for Hydra
    "omegaconf>=2.3.0"    # Added for Hydra's configuration support
]

# Optional: Additional tool configuration
[tool.black]
line-length = 88
target-version = ['py39']

# Pytest configuration options
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov=src --cov-report=xml"
testpaths = [
    "tests"  # Ensure pytest knows where to look for test files
]

